+++ crates/rust-client/src/store/note_record/input_note_record/states/processing_authenticated.rs
@@ -47,7 +47,13 @@ impl NoteStateHandler for ProcessingAuthenticatedNoteState {
         &self,
         nullifier_block_height: BlockNumber,
     ) -> Result<Option<InputNoteState>, NoteRecordError> {
-        Ok(Some(ConsumedExternalNoteState { nullifier_block_height }.into()))
+        Ok(Some(
+            ConsumedExternalNoteState {
+                nullifier_block_height,
+                consumed_tx_order: None,
+            }
+            .into(),
+        ))
     }
 
     fn block_header_received(
@@ -85,6 +91,7 @@ impl NoteStateHandler for ProcessingAuthenticatedNoteState {
                 block_note_root: self.block_note_root,
                 nullifier_block_height: block_height,
                 submission_data: self.submission_data,
+                consumed_tx_order: None,
             }
             .into(),
         ))